split code between exec and spec and removed playgrounds
