--- verus/lcf/src/playground.rs
@@ -1,8 +1,0 @@
-use vstd::prelude::*;
-
-mod string_hash_map;
-mod verusportplayground;
-
-fn main() {
-    // println!("Hello, world!");
-}
